Rename ForcesSCFNTConverter to HedgeSCFNTConverter
Format is also used by Tokyo 2020, no need to have the name seem like it's exclusively meant for Forces
