+++ appveyor.yml
@@ -24,5 +24,5 @@ artifacts:
   name: ForcesGISMEditor
 - path: HedgeTools/HedgeCnvrsEditor/bin/Debug
   name: HedgeCnvrsEditor
-- path: HedgeTools/ForcesSCFNTConverter/bin/Debug
-  name: ForcesSCFNTConverter
+- path: HedgeTools/HedgeSCFNTConverter/bin/Debug
+  name: HedgeSCFNTConverter
